requirements.txt
django
django-ninja
httpx
beautifulsoup4
lxml
celery
redis
eventlet
django-allauth
python-dotenv
psycopg2-binary
django-encrypted-model-fields
dj-database-url